Backend/app/requirements.txt
asgiref==3.9.1
Django
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework-simplejwt
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2

google-auth
django-allauth
social-auth-app-django
requests
python-whois
dnspython==2.6.1
python-magic
celery
redis
channels
channels-redis

